Move `RootBlock` struct into `sp-consensus-spartan`
