+++ crates/sp-consensus-spartan/Cargo.toml
@@ -13,8 +13,17 @@ readme = "README.md"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+parity-scale-codec = { version = "2.0.0", default-features = false }
+ring = { version = "0.16.20", optional = true }
+sha2 = { version = "0.9.8", default-features = false }
+sp-debug-derive = { version = "3.0.0", default-features = false, git = "https://github.com/subspace/substrate", rev = "fa635c968e831110b13f7ff997c4cd3f5e9a3798" }
 spartan-codec = { version = "0.1.0", default-features = false, optional = true }
-ring = { version = "0.16", optional = true }
 
 [features]
-std = ["spartan-codec", "ring"]
+std = [
+    "parity-scale-codec/std",
+    "ring",
+    "sha2/std",
+    "sp-debug-derive/std",
+    "spartan-codec"
+]
